Convert files to TS and add base redux js toolkit setup
