+++ README.md
@@ -1,2 +1,5 @@
-# CountryBucketList
-React Native project with Redux - [Google Play Store](https://play.google.com/store/apps/details?id=com.npdevelopment.countrybucketlist)
+# Country BucketList
+
+React Native project with Expo SDK 48
+
+[Google Play Store](https://play.google.com/store/apps/details?id=com.npdevelopment.countrybucketlist)
