+++ app.json
@@ -3,31 +3,25 @@
     "name": "Country BucketList",
     "slug": "CountryBucketList",
     "privacy": "public",
-    "platforms": [
-      "ios",
-      "android",
-      "web"
-    ],
-    "version": "1.0.3",
+    "platforms": ["ios", "android", "web"],
+    "version": "1.0.4",
     "orientation": "portrait",
-    "icon": "./assets/icon.png",
+    "icon": "./src/assets/icon.png",
     "splash": {
-      "image": "./assets/splash.png",
+      "image": "./src/assets/splash.png",
       "resizeMode": "contain",
       "backgroundColor": "#2196f3"
     },
     "updates": {
       "fallbackToCacheTimeout": 0
     },
-    "assetBundlePatterns": [
-      "**/*"
-    ],
+    "assetBundlePatterns": ["**/*"],
     "ios": {
       "supportsTablet": true
     },
     "android": {
       "package": "com.npdevelopment.countrybucketlist",
-      "versionCode": 4,
+      "versionCode": 5,
       "permissions": ["WAKE_LOCK"]
     }
   }
