+++ package.json
@@ -8,25 +8,32 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@react-native-async-storage/async-storage": "1.17.11",
     "@react-native-community/masked-view": "^0.1.11",
+    "@react-navigation/native": "^6.1.7",
+    "@reduxjs/toolkit": "^1.9.5",
+    "@types/react": "~18.0.27",
     "expo": "^48.0.19",
     "expo-font": "^11.1.1",
     "lodash": "^4.17.21",
     "native-base": "^3.4.28",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-native": "^0.72.1",
-    "react-native-gesture-handler": "^2.12.0",
-    "react-native-screens": "^3.22.0",
+    "react-native": "0.71.8",
+    "react-native-gesture-handler": "~2.9.0",
+    "react-native-safe-area-context": "4.5.0",
+    "react-native-screens": "~3.20.0",
+    "react-native-svg": "13.4.0",
     "react-native-vector-icons": "^9.2.0",
-    "react-native-web": "^0.19.6",
+    "react-native-web": "~0.18.10",
     "react-navigation": "^4.4.4",
     "react-navigation-stack": "^2.10.4",
     "react-redux": "^8.1.1",
+    "react-toastify": "^9.1.3",
     "redux": "^4.2.1",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.4.2"
+    "typescript": "^4.9.4"
   },
   "devDependencies": {
     "babel-preset-expo": "^9.3.2"
