+++ src/data/CountriesFlags.js
@@ -2,1172 +2,1172 @@ const Countries = [
   {
     id: 1,
     country: "Afghanistan",
-    source: require("../../assets/flags/Afghanistan.png"),
-    safe: "No",
+    source: require("../assets/flags/Afghanistan.png"),
+
   },
   {
     id: 2,
     country: "Albania",
-    source: require("../../assets/flags/Albania.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Albania.png"),
+
   },
   {
     id: 3,
     country: "Algeria",
-    source: require("../../assets/flags/Algeria.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Algeria.png"),
+
   },
   {
     id: 4,
     country: "Andorra",
-    source: require("../../assets/flags/Andorra.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Andorra.png"),
+
   },
   {
     id: 5,
     country: "Antigua-and-Barbuda",
-    source: require("../../assets/flags/Antigua-and-Barbuda.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Antigua-and-Barbuda.png"),
+
   },
   {
     id: 6,
     country: "Argentina",
-    source: require("../../assets/flags/Argentina.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Argentina.png"),
+
   },
   {
     id: 7,
     country: "Armenia",
-    source: require("../../assets/flags/Armenia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Armenia.png"),
+
   },
   {
     id: 8,
     country: "Australia",
-    source: require("../../assets/flags/Australia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Australia.png"),
+
   },
   {
     id: 9,
     country: "Austria",
-    source: require("../../assets/flags/Austria.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Austria.png"),
+
   },
   {
     id: 10,
     country: "Azerbaijan",
-    source: require("../../assets/flags/Azerbaijan.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Azerbaijan.png"),
+
   },
   {
     id: 11,
     country: "Bahamas",
-    source: require("../../assets/flags/Bahamas.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Bahamas.png"),
+
   },
   {
     id: 12,
     country: "Bahrain",
-    source: require("../../assets/flags/Bahrain.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Bahrain.png"),
+
   },
   {
     id: 13,
     country: "Bangladesh",
-    source: require("../../assets/flags/Bangladesh.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Bangladesh.png"),
+
   },
   {
     id: 14,
     country: "Barbados",
-    source: require("../../assets/flags/Barbados.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Barbados.png"),
+
   },
   {
     id: 15,
     country: "Belarus",
-    source: require("../../assets/flags/Belarus.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Belarus.png"),
+
   },
   {
     id: 16,
     country: "Belgium",
-    source: require("../../assets/flags/Belgium.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Belgium.png"),
+
   },
   {
     id: 17,
     country: "Belize",
-    source: require("../../assets/flags/Belize.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Belize.png"),
+
   },
   {
     id: 18,
     country: "Benin",
-    source: require("../../assets/flags/Benin.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Benin.png"),
+
   },
   {
     id: 19,
     country: "Bhutan",
-    source: require("../../assets/flags/Bhutan.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Bhutan.png"),
+
   },
   {
     id: 20,
     country: "Bolivia",
-    source: require("../../assets/flags/Bolivia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Bolivia.png"),
+
   },
   {
     id: 21,
     country: "Bosnia-and-Herzegovina",
-    source: require("../../assets/flags/Bosnia-and-Herzegovina.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Bosnia-and-Herzegovina.png"),
+
   },
   {
     id: 22,
     country: "Botswana",
-    source: require("../../assets/flags/Botswana.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Botswana.png"),
+
   },
   {
     id: 23,
     country: "Brazil",
-    source: require("../../assets/flags/Brazil.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Brazil.png"),
+
   },
   {
     id: 24,
     country: "Brunei",
-    source: require("../../assets/flags/Brunei.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Brunei.png"),
+
   },
   {
     id: 25,
     country: "Bulgaria",
-    source: require("../../assets/flags/Bulgaria.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Bulgaria.png"),
+
   },
   {
     id: 26,
     country: "Burkina-Faso",
-    source: require("../../assets/flags/Burkina-Faso.png"),
-    safe: "No",
+    source: require("../assets/flags/Burkina-Faso.png"),
+
   },
   {
     id: 27,
     country: "Burundi",
-    source: require("../../assets/flags/Burundi.png"),
-    safe: "No",
+    source: require("../assets/flags/Burundi.png"),
+
   },
   {
     id: 28,
     country: "Cambodia",
-    source: require("../../assets/flags/Cambodia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Cambodia.png"),
+
   },
   {
     id: 29,
     country: "Cameroon",
-    source: require("../../assets/flags/Cameroon.png"),
-    safe: "No",
+    source: require("../assets/flags/Cameroon.png"),
+
   },
   {
     id: 30,
     country: "Canada",
-    source: require("../../assets/flags/Canada.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Canada.png"),
+
   },
   {
     id: 31,
     country: "Cape-Verde",
-    source: require("../../assets/flags/Cape-Verde.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Cape-Verde.png"),
+
   },
   {
     id: 32,
     country: "Central-African-Republic",
-    source: require("../../assets/flags/Central-African-Republic.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Central-African-Republic.png"),
+
   },
   {
     id: 33,
     country: "Chad",
-    source: require("../../assets/flags/Chad.png"),
-    safe: "No",
+    source: require("../assets/flags/Chad.png"),
+
   },
   {
     id: 34,
     country: "Chile",
-    source: require("../../assets/flags/Chile.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Chile.png"),
+
   },
   {
     id: 35,
     country: "China",
-    source: require("../../assets/flags/China.png"),
-    safe: "Yes",
+    source: require("../assets/flags/China.png"),
+
   },
   {
     id: 36,
     country: "Colombia",
-    source: require("../../assets/flags/Colombia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Colombia.png"),
+
   },
   {
     id: 37,
     country: "Comoros",
-    source: require("../../assets/flags/Comoros.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Comoros.png"),
+
   },
   {
     id: 38,
     country: "Congo-(Democratic)",
-    source: require("../../assets/flags/Congo-(Democratic).png"),
-    safe: "No",
+    source: require("../assets/flags/Congo-(Democratic).png"),
+
   },
   {
     id: 39,
     country: "Congo-(Republic)",
-    source: require("../../assets/flags/Congo-(Republic).png"),
-    safe: "No",
+    source: require("../assets/flags/Congo-(Republic).png"),
+
   },
   {
     id: 40,
     country: "Costa-Rica",
-    source: require("../../assets/flags/Costa-Rica.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Costa-Rica.png"),
+
   },
   {
     id: 41,
     country: "Cote-d'Ivoire",
-    source: require("../../assets/flags/Cote-dIvoire.png"),
-    safe: "No",
+    source: require("../assets/flags/Cote-dIvoire.png"),
+
   },
   {
     id: 42,
     country: "Croatia",
-    source: require("../../assets/flags/Croatia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Croatia.png"),
+
   },
   {
     id: 43,
     country: "Cuba",
-    source: require("../../assets/flags/Cuba.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Cuba.png"),
+
   },
   {
     id: 44,
     country: "Cyprus",
-    source: require("../../assets/flags/Cyprus.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Cyprus.png"),
+
   },
   {
     id: 45,
     country: "Czech-Republic",
-    source: require("../../assets/flags/Czech-Republic.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Czech-Republic.png"),
+
   },
   {
     id: 46,
     country: "Denmark",
-    source: require("../../assets/flags/Denmark.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Denmark.png"),
+
   },
   {
     id: 47,
     country: "Djibouti",
-    source: require("../../assets/flags/Djibouti.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Djibouti.png"),
+
   },
   {
     id: 48,
     country: "Dominica",
-    source: require("../../assets/flags/Dominica.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Dominica.png"),
+
   },
   {
     id: 49,
     country: "Dominican-Republic",
-    source: require("../../assets/flags/Dominican-Republic.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Dominican-Republic.png"),
+
   },
   {
     id: 50,
     country: "East-Timor",
-    source: require("../../assets/flags/East-Timor.png"),
-    safe: "Yes",
+    source: require("../assets/flags/East-Timor.png"),
+
   },
   {
     id: 51,
     country: "Ecuador",
-    source: require("../../assets/flags/Ecuador.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Ecuador.png"),
+
   },
   {
     id: 52,
     country: "Egypt",
-    source: require("../../assets/flags/Egypt.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Egypt.png"),
+
   },
   {
     id: 53,
     country: "El-Salvador",
-    source: require("../../assets/flags/El-Salvador.png"),
-    safe: "Yes",
+    source: require("../assets/flags/El-Salvador.png"),
+
   },
   {
     id: 54,
     country: "Equatorial-Guinea",
-    source: require("../../assets/flags/Equatorial-Guinea.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Equatorial-Guinea.png"),
+
   },
   {
     id: 55,
     country: "Eritrea",
-    source: require("../../assets/flags/Eritrea.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Eritrea.png"),
+
   },
   {
     id: 56,
     country: "Estonia",
-    source: require("../../assets/flags/Estonia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Estonia.png"),
+
   },
   {
     id: 57,
     country: "Ethiopia",
-    source: require("../../assets/flags/Ethiopia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Ethiopia.png"),
+
   },
   {
     id: 58,
     country: "Fiji",
-    source: require("../../assets/flags/Fiji.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Fiji.png"),
+
   },
   {
     id: 59,
     country: "Finland",
-    source: require("../../assets/flags/Finland.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Finland.png"),
+
   },
   {
     id: 60,
     country: "France",
-    source: require("../../assets/flags/France.png"),
-    safe: "Yes",
+    source: require("../assets/flags/France.png"),
+
   },
   {
     id: 61,
     country: "Gabon",
-    source: require("../../assets/flags/Gabon.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Gabon.png"),
+
   },
   {
     id: 62,
     country: "Gambia",
-    source: require("../../assets/flags/Gambia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Gambia.png"),
+
   },
   {
     id: 63,
     country: "Georgia",
-    source: require("../../assets/flags/Georgia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Georgia.png"),
+
   },
   {
     id: 64,
     country: "Germany",
-    source: require("../../assets/flags/Germany.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Germany.png"),
+
   },
   {
     id: 65,
     country: "Ghana",
-    source: require("../../assets/flags/Ghana.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Ghana.png"),
+
   },
   {
     id: 66,
     country: "Greece",
-    source: require("../../assets/flags/Greece.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Greece.png"),
+
   },
   {
     id: 67,
     country: "Grenada",
-    source: require("../../assets/flags/Grenada.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Grenada.png"),
+
   },
   {
     id: 68,
     country: "Guatemala",
-    source: require("../../assets/flags/Guatemala.png"),
-    safe: "No",
+    source: require("../assets/flags/Guatemala.png"),
+
   },
   {
     id: 69,
     country: "Guinea-Bissau",
-    source: require("../../assets/flags/Guinea-Bissau.png"),
-    safe: "No",
+    source: require("../assets/flags/Guinea-Bissau.png"),
+
   },
   {
     id: 70,
     country: "Guinea",
-    source: require("../../assets/flags/Guinea.png"),
-    safe: "No",
+    source: require("../assets/flags/Guinea.png"),
+
   },
   {
     id: 71,
     country: "Guyana",
-    source: require("../../assets/flags/Guyana.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Guyana.png"),
+
   },
   {
     id: 72,
     country: "Haiti",
-    source: require("../../assets/flags/Haiti.png"),
-    safe: "No",
+    source: require("../assets/flags/Haiti.png"),
+
   },
   {
     id: 73,
     country: "Honduras",
-    source: require("../../assets/flags/Honduras.png"),
-    safe: "No",
+    source: require("../assets/flags/Honduras.png"),
+
   },
   {
     id: 74,
     country: "Hungary",
-    source: require("../../assets/flags/Hungary.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Hungary.png"),
+
   },
   {
     id: 75,
     country: "Iceland",
-    source: require("../../assets/flags/Iceland.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Iceland.png"),
+
   },
   {
     id: 76,
     country: "India",
-    source: require("../../assets/flags/India.png"),
-    safe: "Yes",
+    source: require("../assets/flags/India.png"),
+
   },
   {
     id: 77,
     country: "Indonesia",
-    source: require("../../assets/flags/Indonesia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Indonesia.png"),
+
   },
   {
     id: 78,
     country: "Iran",
-    source: require("../../assets/flags/Iran.png"),
-    safe: "No",
+    source: require("../assets/flags/Iran.png"),
+
   },
   {
     id: 79,
     country: "Iraq",
-    source: require("../../assets/flags/Iraq.png"),
-    safe: "No",
+    source: require("../assets/flags/Iraq.png"),
+
   },
   {
     id: 80,
     country: "Ireland",
-    source: require("../../assets/flags/Ireland.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Ireland.png"),
+
   },
   {
     id: 81,
     country: "Israel",
-    source: require("../../assets/flags/Israel.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Israel.png"),
+
   },
   {
     id: 82,
     country: "Italy",
-    source: require("../../assets/flags/Italy.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Italy.png"),
+
   },
   {
     id: 83,
     country: "Jamaica",
-    source: require("../../assets/flags/Jamaica.png"),
-    safe: "No",
+    source: require("../assets/flags/Jamaica.png"),
+
   },
   {
     id: 84,
     country: "Japan",
-    source: require("../../assets/flags/Japan.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Japan.png"),
+
   },
   {
     id: 85,
     country: "Jordan",
-    source: require("../../assets/flags/Jordan.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Jordan.png"),
+
   },
   {
     id: 86,
     country: "Kazakhstan",
-    source: require("../../assets/flags/Kazakhstan.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Kazakhstan.png"),
+
   },
   {
     id: 87,
     country: "Kenya",
-    source: require("../../assets/flags/Kenya.png"),
-    safe: "No",
+    source: require("../assets/flags/Kenya.png"),
+
   },
   {
     id: 88,
     country: "Kiribati",
-    source: require("../../assets/flags/Kiribati.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Kiribati.png"),
+
   },
   {
     id: 89,
     country: "North-Korea",
-    source: require("../../assets/flags/Korea-North.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Korea-North.png"),
+
   },
   {
     id: 90,
     country: "South-Korea",
-    source: require("../../assets/flags/Korea-South.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Korea-South.png"),
+
   },
   {
     id: 91,
     country: "Kosovo",
-    source: require("../../assets/flags/Kosovo.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Kosovo.png"),
+
   },
   {
     id: 92,
     country: "Kuwait",
-    source: require("../../assets/flags/Kuwait.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Kuwait.png"),
+
   },
   {
     id: 93,
     country: "Kyrgyzstan",
-    source: require("../../assets/flags/Kyrgyzstan.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Kyrgyzstan.png"),
+
   },
   {
     id: 94,
     country: "Laos",
-    source: require("../../assets/flags/Laos.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Laos.png"),
+
   },
   {
     id: 95,
     country: "Latvia",
-    source: require("../../assets/flags/Latvia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Latvia.png"),
+
   },
   {
     id: 96,
     country: "Lebanon",
-    source: require("../../assets/flags/Lebanon.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Lebanon.png"),
+
   },
   {
     id: 97,
     country: "Lesotho",
-    source: require("../../assets/flags/Lesotho.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Lesotho.png"),
+
   },
   {
     id: 98,
     country: "Liberia",
-    source: require("../../assets/flags/Liberia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Liberia.png"),
+
   },
   {
     id: 99,
     country: "Libya",
-    source: require("../../assets/flags/Libya.png"),
-    safe: "No",
+    source: require("../assets/flags/Libya.png"),
+
   },
   {
     id: 100,
     country: "Liechtenstein",
-    source: require("../../assets/flags/Liechtenstein.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Liechtenstein.png"),
+
   },
   {
     id: 101,
     country: "Lithuania",
-    source: require("../../assets/flags/Lithuania.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Lithuania.png"),
+
   },
   {
     id: 102,
     country: "Luxembourg",
-    source: require("../../assets/flags/Luxembourg.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Luxembourg.png"),
+
   },
   {
     id: 103,
     country: "Macedonia",
-    source: require("../../assets/flags/Macedonia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Macedonia.png"),
+
   },
   {
     id: 104,
     country: "Madagascar",
-    source: require("../../assets/flags/Madagascar.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Madagascar.png"),
+
   },
   {
     id: 105,
     country: "Malawi",
-    source: require("../../assets/flags/Malawi.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Malawi.png"),
+
   },
   {
     id: 106,
     country: "Malaysia",
-    source: require("../../assets/flags/Malaysia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Malaysia.png"),
+
   },
   {
     id: 107,
     country: "Maldives",
-    source: require("../../assets/flags/Maldives.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Maldives.png"),
+
   },
   {
     id: 108,
     country: "Mali",
-    source: require("../../assets/flags/Mali.png"),
-    safe: "No",
+    source: require("../assets/flags/Mali.png"),
+
   },
   {
     id: 109,
     country: "Malta",
-    source: require("../../assets/flags/Malta.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Malta.png"),
+
   },
   {
     id: 110,
     country: "Marshall-Islands",
-    source: require("../../assets/flags/Marshall-Islands.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Marshall-Islands.png"),
+
   },
   {
     id: 111,
     country: "Mauritania",
-    source: require("../../assets/flags/Mauritania.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Mauritania.png"),
+
   },
   {
     id: 112,
     country: "Mauritius",
-    source: require("../../assets/flags/Mauritius.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Mauritius.png"),
+
   },
   {
     id: 113,
     country: "Mexico",
-    source: require("../../assets/flags/Mexico.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Mexico.png"),
+
   },
   {
     id: 114,
     country: "Micronesia-(Federated)",
-    source: require("../../assets/flags/Micronesia-(Federated).png"),
-    safe: "Yes",
+    source: require("../assets/flags/Micronesia-(Federated).png"),
+
   },
   {
     id: 115,
     country: "Moldova",
-    source: require("../../assets/flags/Moldova.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Moldova.png"),
+
   },
   {
     id: 116,
     country: "Monaco",
-    source: require("../../assets/flags/Monaco.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Monaco.png"),
+
   },
   {
     id: 117,
     country: "Mongolia",
-    source: require("../../assets/flags/Mongolia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Mongolia.png"),
+
   },
   {
     id: 118,
     country: "Montenegro",
-    source: require("../../assets/flags/Montenegro.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Montenegro.png"),
+
   },
   {
     id: 119,
     country: "Morocco",
-    source: require("../../assets/flags/Morocco.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Morocco.png"),
+
   },
   {
     id: 120,
     country: "Mozambique",
-    source: require("../../assets/flags/Mozambique.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Mozambique.png"),
+
   },
   {
     id: 121,
     country: "Myanmar",
-    source: require("../../assets/flags/Myanmar.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Myanmar.png"),
+
   },
   {
     id: 122,
     country: "Namibia",
-    source: require("../../assets/flags/Namibia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Namibia.png"),
+
   },
   {
     id: 123,
     country: "Nauru",
-    source: require("../../assets/flags/Nauru.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Nauru.png"),
+
   },
   {
     id: 124,
     country: "Nepal",
-    source: require("../../assets/flags/Nepal.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Nepal.png"),
+
   },
   {
     id: 125,
     country: "Netherlands",
-    source: require("../../assets/flags/Netherlands.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Netherlands.png"),
+
   },
   {
     id: 126,
     country: "New-Zealand",
-    source: require("../../assets/flags/New-Zealand.png"),
-    safe: "Yes",
+    source: require("../assets/flags/New-Zealand.png"),
+
   },
   {
     id: 127,
     country: "Nicaragua",
-    source: require("../../assets/flags/Nicaragua.png"),
-    safe: "No",
+    source: require("../assets/flags/Nicaragua.png"),
+
   },
   {
     id: 128,
     country: "Niger",
-    source: require("../../assets/flags/Niger.png"),
-    safe: "No",
+    source: require("../assets/flags/Niger.png"),
+
   },
   {
     id: 129,
     country: "Nigeria",
-    source: require("../../assets/flags/Nigeria.png"),
-    safe: "No",
+    source: require("../assets/flags/Nigeria.png"),
+
   },
   {
     id: 130,
     country: "Norway",
-    source: require("../../assets/flags/Norway.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Norway.png"),
+
   },
   {
     id: 131,
     country: "Oman",
-    source: require("../../assets/flags/Oman.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Oman.png"),
+
   },
   {
     id: 132,
     country: "Pakistan",
-    source: require("../../assets/flags/Pakistan.png"),
-    safe: "No",
+    source: require("../assets/flags/Pakistan.png"),
+
   },
   {
     id: 133,
     country: "Palau",
-    source: require("../../assets/flags/Palau.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Palau.png"),
+
   },
   {
     id: 134,
     country: "Panama",
-    source: require("../../assets/flags/Panama.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Panama.png"),
+
   },
   {
     id: 135,
     country: "Papua-New-Guinea",
-    source: require("../../assets/flags/Papua-New-Guinea.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Papua-New-Guinea.png"),
+
   },
   {
     id: 136,
     country: "Paraguay",
-    source: require("../../assets/flags/Paraguay.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Paraguay.png"),
+
   },
   {
     id: 137,
     country: "Peru",
-    source: require("../../assets/flags/Peru.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Peru.png"),
+
   },
   {
     id: 138,
     country: "Philippines",
-    source: require("../../assets/flags/Philippines.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Philippines.png"),
+
   },
   {
     id: 139,
     country: "Poland",
-    source: require("../../assets/flags/Poland.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Poland.png"),
+
   },
   {
     id: 140,
     country: "Portugal",
-    source: require("../../assets/flags/Portugal.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Portugal.png"),
+
   },
   {
     id: 141,
     country: "Qatar",
-    source: require("../../assets/flags/Qatar.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Qatar.png"),
+
   },
   {
     id: 142,
     country: "Romania",
-    source: require("../../assets/flags/Romania.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Romania.png"),
+
   },
   {
     id: 143,
     country: "Russia",
-    source: require("../../assets/flags/Russia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Russia.png"),
+
   },
   {
     id: 144,
     country: "Rwanda",
-    source: require("../../assets/flags/Rwanda.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Rwanda.png"),
+
   },
   {
     id: 145,
     country: "Saint-Kitts-and-Nevis",
-    source: require("../../assets/flags/Saint-Kitts-and-Nevis.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Saint-Kitts-and-Nevis.png"),
+
   },
   {
     id: 146,
     country: "Saint-Lucia",
-    source: require("../../assets/flags/Saint-Lucia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Saint-Lucia.png"),
+
   },
   {
     id: 147,
     country: "Saint-Vincent-and-the-Grenadines",
-    source: require("../../assets/flags/Saint-Vincent-and-the-Grenadines.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Saint-Vincent-and-the-Grenadines.png"),
+
   },
   {
     id: 148,
     country: "Samoa",
-    source: require("../../assets/flags/Samoa.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Samoa.png"),
+
   },
   {
     id: 149,
     country: "San-Marino",
-    source: require("../../assets/flags/San-Marino.png"),
-    safe: "Yes",
+    source: require("../assets/flags/San-Marino.png"),
+
   },
   {
     id: 150,
     country: "Sao-Tome-and-Principe",
-    source: require("../../assets/flags/Sao-Tome-and-Principe.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Sao-Tome-and-Principe.png"),
+
   },
   {
     id: 151,
     country: "Saudi-Arabia",
-    source: require("../../assets/flags/Saudi-Arabia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Saudi-Arabia.png"),
+
   },
   {
     id: 152,
     country: "Senegal",
-    source: require("../../assets/flags/Senegal.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Senegal.png"),
+
   },
   {
     id: 153,
     country: "Serbia",
-    source: require("../../assets/flags/Serbia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Serbia.png"),
+
   },
   {
     id: 154,
     country: "Seychelles",
-    source: require("../../assets/flags/Seychelles.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Seychelles.png"),
+
   },
   {
     id: 155,
     country: "Sierra-Leone",
-    source: require("../../assets/flags/Sierra-Leone.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Sierra-Leone.png"),
+
   },
   {
     id: 156,
     country: "Singapore",
-    source: require("../../assets/flags/Singapore.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Singapore.png"),
+
   },
   {
     id: 157,
     country: "Slovakia",
-    source: require("../../assets/flags/Slovakia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Slovakia.png"),
+
   },
   {
     id: 158,
     country: "Slovenia",
-    source: require("../../assets/flags/Slovenia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Slovenia.png"),
+
   },
   {
     id: 159,
     country: "Solomon-Islands",
-    source: require("../../assets/flags/Solomon-Islands.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Solomon-Islands.png"),
+
   },
   {
     id: 160,
     country: "Somalia",
-    source: require("../../assets/flags/Somalia.png"),
-    safe: "No",
+    source: require("../assets/flags/Somalia.png"),
+
   },
   {
     id: 161,
     country: "South-Africa",
-    source: require("../../assets/flags/South-Africa.png"),
-    safe: "Yes",
+    source: require("../assets/flags/South-Africa.png"),
+
   },
   {
     id: 162,
     country: "South-Sudan",
-    source: require("../../assets/flags/South-Sudan.png"),
-    safe: "No",
+    source: require("../assets/flags/South-Sudan.png"),
+
   },
   {
     id: 163,
     country: "Spain",
-    source: require("../../assets/flags/Spain.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Spain.png"),
+
   },
   {
     id: 164,
     country: "Sri-Lanka",
-    source: require("../../assets/flags/Sri-Lanka.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Sri-Lanka.png"),
+
   },
   {
     id: 165,
     country: "Sudan",
-    source: require("../../assets/flags/Sudan.png"),
-    safe: "No",
+    source: require("../assets/flags/Sudan.png"),
+
   },
   {
     id: 166,
     country: "Suriname",
-    source: require("../../assets/flags/Suriname.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Suriname.png"),
+
   },
   {
     id: 167,
     country: "Swaziland",
-    source: require("../../assets/flags/Swaziland.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Swaziland.png"),
+
   },
   {
     id: 168,
     country: "Sweden",
-    source: require("../../assets/flags/Sweden.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Sweden.png"),
+
   },
   {
     id: 169,
     country: "Switzerland",
-    source: require("../../assets/flags/Switzerland.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Switzerland.png"),
+
   },
   {
     id: 170,
     country: "Syria",
-    source: require("../../assets/flags/Syria.png"),
-    safe: "No",
+    source: require("../assets/flags/Syria.png"),
+
   },
   {
     id: 171,
     country: "Taiwan",
-    source: require("../../assets/flags/Taiwan.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Taiwan.png"),
+
   },
   {
     id: 172,
     country: "Tajikistan",
-    source: require("../../assets/flags/Tajikistan.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Tajikistan.png"),
+
   },
   {
     id: 173,
     country: "Tanzania",
-    source: require("../../assets/flags/Tanzania.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Tanzania.png"),
+
   },
   {
     id: 174,
     country: "Thailand",
-    source: require("../../assets/flags/Thailand.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Thailand.png"),
+
   },
   {
     id: 175,
     country: "Togo",
-    source: require("../../assets/flags/Togo.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Togo.png"),
+
   },
   {
     id: 176,
     country: "Tonga",
-    source: require("../../assets/flags/Tonga.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Tonga.png"),
+
   },
   {
     id: 177,
     country: "Trinidad-and-Tobago",
-    source: require("../../assets/flags/Trinidad-and-Tobago.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Trinidad-and-Tobago.png"),
+
   },
   {
     id: 178,
     country: "Tunisia",
-    source: require("../../assets/flags/Tunisia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Tunisia.png"),
+
   },
   {
     id: 179,
     country: "Turkey",
-    source: require("../../assets/flags/Turkey.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Turkey.png"),
+
   },
   {
     id: 180,
     country: "Turkmenistan",
-    source: require("../../assets/flags/Turkmenistan.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Turkmenistan.png"),
+
   },
   {
     id: 181,
     country: "Tuvalu",
-    source: require("../../assets/flags/Tuvalu.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Tuvalu.png"),
+
   },
   {
     id: 182,
     country: "Uganda",
-    source: require("../../assets/flags/Uganda.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Uganda.png"),
+
   },
   {
     id: 183,
     country: "Ukraine",
-    source: require("../../assets/flags/Ukraine.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Ukraine.png"),
+
   },
   {
     id: 184,
     country: "United-Arab-Emirates",
-    source: require("../../assets/flags/United-Arab-Emirates.png"),
-    safe: "Yes",
+    source: require("../assets/flags/United-Arab-Emirates.png"),
+
   },
   {
     id: 185,
     country: "United-Kingdom",
-    source: require("../../assets/flags/United-Kingdom.png"),
-    safe: "Yes",
+    source: require("../assets/flags/United-Kingdom.png"),
+
   },
   {
     id: 186,
     country: "United-States-of-America",
-    source: require("../../assets/flags/United-States-of-America.png"),
-    safe: "Yes",
+    source: require("../assets/flags/United-States-of-America.png"),
+
   },
   {
     id: 187,
     country: "Uruguay",
-    source: require("../../assets/flags/Uruguay.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Uruguay.png"),
+
   },
   {
     id: 188,
     country: "Uzbekistan",
-    source: require("../../assets/flags/Uzbekistan.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Uzbekistan.png"),
+
   },
   {
     id: 189,
     country: "Vanuatu",
-    source: require("../../assets/flags/Vanuatu.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Vanuatu.png"),
+
   },
   {
     id: 190,
     country: "Vatican-City",
-    source: require("../../assets/flags/Vatican-City.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Vatican-City.png"),
+
   },
   {
     id: 191,
     country: "Venezuela",
-    source: require("../../assets/flags/Venezuela.png"),
-    safe: "No",
+    source: require("../assets/flags/Venezuela.png"),
+
   },
   {
     id: 192,
     country: "Vietnam",
-    source: require("../../assets/flags/Vietnam.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Vietnam.png"),
+
   },
   {
     id: 193,
     country: "Yemen",
-    source: require("../../assets/flags/Yemen.png"),
-    safe: "No",
+    source: require("../assets/flags/Yemen.png"),
+
   },
   {
     id: 194,
     country: "Zambia",
-    source: require("../../assets/flags/Zambia.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Zambia.png"),
+
   },
   {
     id: 195,
     country: "Zimbabwe",
-    source: require("../../assets/flags/Zimbabwe.png"),
-    safe: "Yes",
+    source: require("../assets/flags/Zimbabwe.png"),
+
   },
 ];
