+++ src/interfaces/bucketlist.interface.ts
@@ -1,0 +1,5 @@
+export interface IBucketList {
+  countryId: number;
+  ideas: string[];
+  achieved: number[];
+}
