+++ src/interfaces/country.interface.ts
@@ -1,0 +1,4 @@
+export interface ICountry {
+  id: number;
+  name: string;
+}
