--- src/reducers/messageReducer.js
@@ -1,13 +1,0 @@
-import { ERROR_MESSAGE } from "../actions/types";
-
-export default (state = [], action) => {
-  switch (action.type) {
-    case ERROR_MESSAGE:
-      return {
-        ...state,
-        allErrors: [...state.allErrors, action.payload]
-      };
-    default:
-      return state;
-  }
-};
